requirements.txt
Django==4.2.2
Pillow==9.2.0
gunicorn==20.1.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
djangorestframework==3.14.0
sqlparse==0.4.4